Utilities moved outside the core, into a higher abstraction layer in the wrapper module
